requirements.txt
appdirs==1.4.0
awesome-slugify==1.6.5
bleach==2.0.0
click==6.7
Flask==0.12
Flask-SQLAlchemy==2.1
Flask-Uploads==0.2.1
Flask-WTF==0.14.2
gunicorn==19.6.0
html5lib==0.999999999
itsdangerous==0.24
Jinja2==2.9.5
MarkupSafe==0.23
packaging==16.8
psycopg2==2.6.2
pyparsing==2.1.10
regex==2017.2.8
six==1.10.0
SQLAlchemy==1.1.5
Unidecode==0.4.20
webencodings==0.5
Werkzeug==0.11.15
WTForms==2.1
